Refactored resuable code
