+++ run_experiments.py
@@ -2,7 +2,7 @@ import importlib
 import json
 from glob import glob
 
-from bitnet.config import ProjectConfig, ExperimentConfig
+from bitnet.experiments.config import ProjectConfig, ExperimentConfig
 
 
 def run_experiment(module_name: str) -> dict:
